+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/submodel/domain/repository/SubmodelServerRepository.java
@@ -20,7 +20,7 @@
 package org.eclipse.tractusx.traceability.submodel.domain.repository;
 
 public interface SubmodelServerRepository {
-    void saveSubmodel(String submodelId, String submodel);
+    String saveSubmodel(String submodel);
 
     String getSubmodel(String submodelId);
 }