+++ CHANGELOG.md
@@ -10,6 +10,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Added
 - #350 Added properties for provisioning to submodel / dtr service
 
+### Added
+- #62 - Added configurable assert refresh cron job with default values
 
 ## [13.0.2-cfx-4]